requirements.txt
aiogram==3.13.0
ccxt
pandas
python-dotenv